+++ internal/compiler/lint/rules/a11y/noAriaUnsupportedElements.test.rjson
@@ -1,0 +1,42 @@
+[
+{
+filename: "file.tsx"
+invalid: [
+	"<meta charset=\"UTF-8\" aria-hidden=\"false\" />"
+	"<meta charset=\"UTF-8\" role=\"meta\" />"
+	"<html aria-required=\"true\" />"
+	"<html role=\"html\"></html>"
+	"<script aria-label=\"script\"></script>"
+	"<script role=\"script\"></script>"
+	"<style aria-labelledby></style>"
+	"<style role=\"style\"></style>"
+]
+valid: [
+	"<meta charset=\"UTF-8\" />"
+	"<html maria=\"text\"></html>"
+	"<script></script>"
+	"<style parole></style>"
+]
+
+},
+{
+filename: "file.html"
+invalid: [
+	"<meta charset=\"UTF-8\" aria-hidden=\"false\" />"
+	"<meta charset=\"UTF-8\" role=\"meta\" />"
+	"<html aria-required=\"true\" />"
+	"<html role=\"html\"></html>"
+	"<script aria-label=\"script\"></script>"
+	"<script role=\"script\"></script>"
+	"<style aria-labelledby=\"id\"></style>"
+	"<style role=\"style\"></style>"
+]
+valid: [
+	"<meta charset=\"UTF-8\" />"
+	"<html maria=\"text\"></html>"
+	"<script></script>"
+	"<style parole></style>"
+]
+
+}
+]
