--- internal/compiler/lint/rules/jsx-a11y/noAriaUnsupportedElements.test.md
@@ -1,285 +1,0 @@
-# `harness.test.ts`
-
-**DO NOT MODIFY**. This file has been autogenerated. Run `rome test internal/compiler/lint/rules/harness.test.ts --update-snapshots` to update.
-
-## `jsx-a11y/noAriaUnsupportedElements`
-
-### `0`
-
-```
-
- lint/jsx-a11y/noAriaUnsupportedElements/reject/1/file.tsx:1 lint/jsx-a11y/noAriaUnsupportedElements
-  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-  ✖ Avoid the role attribute and aria-* attributes when using meta, html, script, and style
-    elements.
-
-    <meta charset='UTF-8' aria-hidden='false' />
-    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-  ℹ Using roles on elements that do not support them can cause issues with screen readers.
-
-━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-✖ Found 1 problem
-
-```
-
-### `0: formatted`
-
-```tsx
-<meta charset="UTF-8" aria-hidden="false" />;
-
-```
-
-### `1`
-
-```
-
- lint/jsx-a11y/noAriaUnsupportedElements/reject/2/file.tsx:1 lint/jsx-a11y/noAriaUnsupportedElements
-  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-  ✖ Avoid the role attribute and aria-* attributes when using meta, html, script, and style
-    elements.
-
-    <meta charset='UTF-8' role='meta' />
-    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-  ℹ Using roles on elements that do not support them can cause issues with screen readers.
-
-━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-✖ Found 1 problem
-
-```
-
-### `1: formatted`
-
-```tsx
-<meta charset="UTF-8" role="meta" />;
-
-```
-
-### `2`
-
-```
-
- lint/jsx-a11y/noAriaUnsupportedElements/reject/3/file.tsx:1 lint/jsx-a11y/noAriaUnsupportedElements
-  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-  ✖ Avoid the role attribute and aria-* attributes when using meta, html, script, and style
-    elements.
-
-    <html aria-required='true' />
-    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-  ℹ Using roles on elements that do not support them can cause issues with screen readers.
-
-━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-✖ Found 1 problem
-
-```
-
-### `2: formatted`
-
-```tsx
-<html aria-required="true" />;
-
-```
-
-### `3`
-
-```
-
- lint/jsx-a11y/noAriaUnsupportedElements/reject/4/file.tsx:1 lint/jsx-a11y/noAriaUnsupportedElements
-  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-  ✖ Avoid the role attribute and aria-* attributes when using meta, html, script, and style
-    elements.
-
-    <html role='html'></html>
-    ^^^^^^^^^^^^^^^^^^^^^^^^^
-
-  ℹ Using roles on elements that do not support them can cause issues with screen readers.
-
-━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-✖ Found 1 problem
-
-```
-
-### `3: formatted`
-
-```tsx
-<html role="html" />;
-
-```
-
-### `4`
-
-```
-
- lint/jsx-a11y/noAriaUnsupportedElements/reject/5/file.tsx:1 lint/jsx-a11y/noAriaUnsupportedElements
-  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-  ✖ Avoid the role attribute and aria-* attributes when using meta, html, script, and style
-    elements.
-
-    <script aria-label='script'></script>
-    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-  ℹ Using roles on elements that do not support them can cause issues with screen readers.
-
-━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-✖ Found 1 problem
-
-```
-
-### `4: formatted`
-
-```tsx
-<script aria-label="script" />;
-
-```
-
-### `5`
-
-```
-
- lint/jsx-a11y/noAriaUnsupportedElements/reject/6/file.tsx:1 lint/jsx-a11y/noAriaUnsupportedElements
-  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-  ✖ Avoid the role attribute and aria-* attributes when using meta, html, script, and style
-    elements.
-
-    <script role='script'></script>
-    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-  ℹ Using roles on elements that do not support them can cause issues with screen readers.
-
-━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-✖ Found 1 problem
-
-```
-
-### `5: formatted`
-
-```tsx
-<script role="script" />;
-
-```
-
-### `6`
-
-```
-
- lint/jsx-a11y/noAriaUnsupportedElements/reject/7/file.tsx:1 lint/jsx-a11y/noAriaUnsupportedElements
-  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-  ✖ Avoid the role attribute and aria-* attributes when using meta, html, script, and style
-    elements.
-
-    <style aria-labelledby></style>
-    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-  ℹ Using roles on elements that do not support them can cause issues with screen readers.
-
-━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-✖ Found 1 problem
-
-```
-
-### `6: formatted`
-
-```tsx
-<style aria-labelledby={true} />;
-
-```
-
-### `7`
-
-```
-
- lint/jsx-a11y/noAriaUnsupportedElements/reject/8/file.tsx:1 lint/jsx-a11y/noAriaUnsupportedElements
-  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-  ✖ Avoid the role attribute and aria-* attributes when using meta, html, script, and style
-    elements.
-
-    <style role='style'></style>
-    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-  ℹ Using roles on elements that do not support them can cause issues with screen readers.
-
-━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-✖ Found 1 problem
-
-```
-
-### `7: formatted`
-
-```tsx
-<style role="style" />;
-
-```
-
-### `8`
-
-```
-✔ No known problems!
-
-```
-
-### `8: formatted`
-
-```tsx
-<meta charset="UTF-8" />;
-
-```
-
-### `9`
-
-```
-✔ No known problems!
-
-```
-
-### `9: formatted`
-
-```tsx
-<html maria="text" />;
-
-```
-
-### `10`
-
-```
-✔ No known problems!
-
-```
-
-### `10: formatted`
-
-```tsx
-<script />;
-
-```
-
-### `11`
-
-```
-✔ No known problems!
-
-```
-
-### `11: formatted`
-
-```tsx
-<style parole={true} />;
-
-```
